first attempt at integrating deck into callbreak
Initially I put the deck into a different crate because
I thought other games might be able to use the same deck.
However Rank and Suit implement PartialEq, PartialOrd
which are only relevant for callbreak game. currently
I am thinking it is best to have it all in one crate.